production_delete

+++ src/reducks/productions/operations/delete/images.js
@@ -1,8 +1,9 @@
 import { deleteObject, ref } from "firebase/storage";
 import { storage } from "../../../../firebase";
 
-const images=({id})=>{
+const images=(id)=>{
     return async(dispatch,setState)=>{
+
         const state=setState();
         const productions =state.productions.productions;
         const user = state.user
@@ -10,10 +11,13 @@ const images=({id})=>{
         productions.forEach((value,num)=>{
             if(id===value.id){i=num}
         })
-        productions[i].images.forEach((value,i)=>{
-            const productionStorageRef=ref(storage,"images/productions/"+productions.id+"/production-"+i+".jpg")
-            deleteObject(productionStorageRef)
-        })
+        if(productions[i].images){
+            productions[i].images.forEach((value,i)=>{
+                const productionStorageRef=ref(storage,"images/productions/"+id+"/production-"+i+".jpg")
+                deleteObject(productionStorageRef)
+            })
+        }
+
     }
 }
